requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
gspread==6.1.2
google-auth==2.34.0
openai-whisper==20231117
